Refactor shared logic into common module and add linting docs
