+++ README.md
@@ -35,3 +35,12 @@ Edit `config.yaml` to adjust limits and model settings. Key options include:
 - `openai.draft_max_tokens` – token limit for reply generation.
 - `openai.classify_max_tokens` – token limit for the classification step.
 - `limits.max_drafts` – maximum number of drafts created per run.
+
+### Development
+
+Install `flake8` and run it before committing changes:
+
+```bash
+pip install flake8
+flake8
+```
